+++ R/create_indiedown_package.R
@@ -38,9 +38,14 @@ create_indiedown_package <- function(path, overwrite = FALSE) {
   files <- c(
     "inst/rmarkdown/templates/report/skeleton/skeleton.Rmd",
     "R/indiedown_pdf_document.R",
+    "R/use_quarto.R",
     "man/mypackage.Rd",
+    "man/use_mypackage_quarto.Rd",
     "DESCRIPTION",
-    "NAMESPACE"
+    "NAMESPACE",
+    "inst/quarto/_extensions/mypackage/mypackage/_extension.yml",
+    "inst/quarto/_extensions/mypackage/mypackage/preamble.tex",
+    "inst/quarto/examples/example.qmd"
   )
 
   gsub_in_file(
@@ -49,10 +54,31 @@ create_indiedown_package <- function(path, overwrite = FALSE) {
     file = files
   )
 
-  file.rename(
-    "man/mypackage.Rd",
-    file.path("man", paste0(pkg_name, ".Rd"))
-  )
+  if (pkg_name != "mypackage") {
+    file.rename(
+      "man/mypackage.Rd",
+      file.path("man", paste0(pkg_name, ".Rd"))
+    )
+    file.rename(
+      "man/use_mypackage_quarto.Rd",
+      file.path("man", paste0("use_", pkg_name, "_quarto.Rd"))
+    )
+  }
+
+  # Rename the Quarto extension directories from mypackage → pkg_name.
+  # Order matters: rename the inner directory first, then the outer.
+  if (pkg_name != "mypackage") {
+    inner_old <- file.path("inst/quarto/_extensions", "mypackage", "mypackage")
+    inner_new <- file.path("inst/quarto/_extensions", "mypackage", pkg_name)
+    if (fs::dir_exists(inner_old)) {
+      fs::file_move(inner_old, inner_new)
+    }
+    outer_old <- file.path("inst/quarto/_extensions", "mypackage")
+    outer_new <- file.path("inst/quarto/_extensions", pkg_name)
+    if (fs::dir_exists(outer_old)) {
+      fs::file_move(outer_old, outer_new)
+    }
+  }
 
   # Can't use .here as part of the template due to the leading dot
   writeLines(character(), ".here")